ajout pop up de confirmation

+++ SAE_IHM/src/main/java/fr/univ_orleans/iut45/controleur/SupprimerBoiteControleur.java
@@ -5,11 +5,14 @@ import fr.univ_orleans.iut45.modele.BoiteSimple;
 import fr.univ_orleans.iut45.vue.Vue;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.VBox;
 import java.util.List;
+import java.util.Optional;
 
 public class SupprimerBoiteControleur {
 
@@ -104,13 +107,22 @@ public class SupprimerBoiteControleur {
         if (boiteSelectionnee != null) {
             try {
                 BoiteBD boiteBD = new BoiteBD(vue.getConnexionMySQL());
-                boiteBD.supprimerBoitePartout(boiteSelectionnee.getNumBoite());
+                Optional<ButtonType> reponse = popUpConfirmation().showAndWait(); 
+                if (reponse.isPresent() && reponse.get().equals(ButtonType.YES)){
+                    System.out.println("Ok !");
+                    boiteBD.supprimerBoitePartout(boiteSelectionnee.getNumBoite());
                 
                 labelMessage.setText("La boîte " + boiteSelectionnee.getNumBoite() + " a été supprimée avec succès !");
                 labelMessage.setStyle("-fx-text-fill: #1A6B3C; -fx-font-weight: bold;"); 
                 
                 champRecherche.clear();
                 cacherCarte();
+                }
+                else{
+                    System.out.println("D'ac !");
+                    
+                }
+                
                 
             } catch (Exception e) {
                 labelMessage.setText("Erreur lors de la suppression.");
@@ -119,4 +131,12 @@ public class SupprimerBoiteControleur {
             }
         }
     }
+
+
+    public Alert popUpConfirmation(){
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION,"Vous allez supprimer une boite de la collection cette action sera irréversible", ButtonType.YES, ButtonType.NO);
+        alert.setTitle("Attention");
+        return alert;
+    }
+
 }
